refactor: move api::tests::helpers to test::lib
